refactor(tests): fold the stress tier onto the shared XRAS helpers
## Summary

`61370a7` consolidated the XRAS API auth fixtures for `tests/api/` and left
`tests/stress/conftest.py` holding a **third** copy of the same four things —
`XRAS_PW`, `auth_headers`, `xras_client`, `_reset_db_key_cache`. Leaving it is how
the copies drift back, so it goes now rather than after the tier grows.

Also collapsed, all of it duplication the pre-cutover tidy-up flagged but did not
reach:

- `_post` was byte-for-byte identical in `test_parking_is_explained.py:27` and
  `test_repeat_and_unsampled.py:26` → `post_action` in the conftest.
- `PATH = '/api/xras/v1/actions'` was spelled in all three modules → `ACTIONS_PATH`.
- `any_active_project` and `snapshot_project` were the same fixture under two names,
  one per module → one `snapshot_project`, keeping the better of the two docstrings
  (the one explaining that a factory project is invisible to the route and silently
  turns a scenario into a different scenario).
- `no_handlers` deleted — no test requested it. A leftover from the two modules whose
  stale `.pyc` files are still in `__pycache__`.

The stress `xras_client` installed only the `samuel` key where the shared one installs
`samuel` + `nobody`; a superset is harmless, so the shared one wins.

### The wrinkle worth recording

Importing `xras_client` alone fails with `fixture 'xras_keys' not found` — pytest
resolves a fixture's own dependencies by name in the **requesting** module's
namespace, so a dependent fixture has to be imported alongside it. It surfaces at
setup rather than import, so it errors every test in the tier rather than failing
loudly at collection. Noted on the fixture itself.

### Test Results

`pytest -m stress -n 0` → **17 passed**, with no test-body edits beyond the mechanical
rename — which is the proof the move was pure.
