+++ tests/stress/conftest.py
@@ -9,19 +9,27 @@ Mirrors ``tests/perf/conftest.py``: a JSON manifest read once at import, exposed
 a fixture, with the same "declare the expectation next to the number" convention.
 """
 
-import base64
 import json
 from pathlib import Path
 
-import bcrypt
 import pytest
 
 from xras_audit import action_log  # noqa: F401  — shared with tests/api/
-from webapp.utils import api_auth
+from xras_helpers import (  # noqa: F401  — pytest resolves fixtures by name
+    reset_db_key_cache,
+    xras_auth as auth_headers,
+    xras_client,
+    xras_keys,          # `xras_client` requests it; importing one without the
+)                       # other fails at setup, not at import
+
+
+
 
 _SCENARIOS_PATH = Path(__file__).parent / 'scenarios.json'
 
-XRAS_PW = 'xras-stress-pw'
+#: The one path every route-driven scenario posts to. Defined here rather than
+#: re-declared per module, which is how it came to be spelled three times.
+ACTIONS_PATH = '/api/xras/v1/actions'
 
 
 def _load_scenarios():
@@ -49,30 +57,37 @@ def scenario(request):
 
 
 # ---------------------------------------------------------------------------
-# Auth — same shape as tests/api/test_xras_access.py, which is the precedent.
-# A `make_api_credentials` row is invisible to an HTTP request (routes read
-# db.session on a separate connection), so the DB-key loader is patched instead.
+# Posting
 # ---------------------------------------------------------------------------
 
-@pytest.fixture(autouse=True)
-def _reset_db_key_cache():
-    api_auth._DB_KEY_CACHE.update(at=None, map={})
-    yield
-    api_auth._DB_KEY_CACHE.update(at=None, map={})
+def post_action(client, payload):
+    """POST one action payload. Was defined byte-for-byte twice, in two modules."""
+    return client.post(ACTIONS_PATH, data=json.dumps(payload),
+                       content_type='application/json', headers=auth_headers())
 
 
 @pytest.fixture
-def xras_client(client, monkeypatch):
-    hashed = bcrypt.hashpw(XRAS_PW.encode(), bcrypt.gensalt(rounds=4)).decode()
-    monkeypatch.setattr(
-        api_auth, '_get_db_api_keys',
-        lambda: {'samuel': {'hash': hashed, 'roles': ['ROLE_XRAS']}})
-    return client
+def snapshot_project(app):
+    """A committed snapshot project.
 
+    ⚠️ Route tests cannot use factories. The route reads Flask-SQLAlchemy's
+    ``db.session`` on its own connection, which sees only **committed** rows, so a
+    factory-made project is invisible to it and every dispatch would park as "no
+    service matches" — silently turning a scenario into a different scenario.
 
-def auth_headers(username='samuel'):
-    token = base64.b64encode(f'{username}:{XRAS_PW}'.encode()).decode('ascii')
-    return {'Authorization': f'Basic {token}'}
+    Was two fixtures with identical bodies under two names, one per module —
+    ``snapshot_project`` and ``any_active_project``. One name now.
+    """
+    from sqlalchemy.orm import Session
+
+    from sam.projects.projects import Project
+    from webapp.extensions import db
+
+    with app.app_context(), Session(db.engine) as session:
+        project = session.query(Project).filter(Project.is_active).first()
+        assert project is not None, 'the snapshot has no active project'
+        session.expunge(project)
+        return project
 
 
 @pytest.fixture
@@ -89,13 +104,6 @@ def dispatching(app):
         app.config.pop('XRAS_ACTIONS_ENABLED', None)
         dispatch._HANDLERS.clear()
         dispatch._HANDLERS.update(saved)
-
-
-@pytest.fixture
-def no_handlers(dispatching):
-    """Capture off **and** an empty registry — the 'nothing is registered' park."""
-    dispatching._HANDLERS.clear()
-    return dispatching
 
 
 @pytest.fixture
